requirements.txt
fastapi
motor
pymongo
python-jose
passlib[bcrypt]
pydantic[email]
python-dotenv
bcrypt==4.0.1